some refactoring to adhere to coding guidelines; Made it look like the UI draft
