+++ src/Data/Services/ItemsService.cs
@@ -4,13 +4,14 @@ public class ItemsService(ApplicationDbContext context)
 {    
     public async Task CreateItemAsync(long projectID, string name)
         {
-            ItemDefinition Item = new ItemDefinition
+            ItemDefinition newItem = new ItemDefinition
             {
                 ItemName = name,
-                Project = await context.Projects.FindAsync(projectID) ?? throw new Exception("Invalid Project ID for Item Creation")
+                Project = await context.Projects.FindAsync(projectID) 
+                ?? throw new Exception("Invalid Project ID for Item Creation")
             };
 
-            context.ItemDefinitions.Add(Item);
+            context.ItemDefinitions.Add(newItem);
             await context.SaveChangesAsync();
 
             return;
